MGS: Expect a hubris archive for SP updates
Previously, faux-mgs expected a hubris archive, but MGS proper expected
a raw image. Now both expect a hubris archive.
