+++ gateway-sp-comms/src/error.rs
@@ -35,6 +35,12 @@ pub enum UpdateError {
     ImageEmpty,
     #[error("update image is too large")]
     ImageTooLarge,
+    #[error("failed to parse SP update as a zip file: {0}")]
+    SpUpdateNotZip(zip::result::ZipError),
+    #[error("failed to find `{path}` within SP update: {err}")]
+    SpUpdateFileNotFound { path: String, err: zip::result::ZipError },
+    #[error("failed to decompress `{path}` within SP update: {err}")]
+    SpUpdateDecompressionFailed { path: String, err: io::Error },
     #[error("failed to send update message to SP: {0}")]
     Communication(#[from] SpCommunicationError),
 }
